+++ nodes/knxUltimateHueLight.js
@@ -3,11 +3,10 @@
 /* eslint-disable max-len */
 /* eslint-disable no-lonely-if */
 const cloneDeep = require("lodash/cloneDeep");
+const dptlib = require("../KNXEngine/src/dptlib");
+const hueColorConverter = require("./utils/colorManipulators/hueColorConverter");
 
 module.exports = function (RED) {
-  const dptlib = require("../KNXEngine/src/dptlib");
-  const hueColorConverter = require("./utils/hueColorConverter");
-
   function knxUltimateHueLight(config) {
     RED.nodes.createNode(this, config);
     const node = this;
@@ -61,8 +60,10 @@ module.exports = function (RED) {
     config.colorAtSwitchOnNightTime = (config.colorAtSwitchOnNightTime === '' || config.colorAtSwitchOnNightTime === undefined) ? '{ "kelvin":2700, "brightness":20 }' : config.colorAtSwitchOnNightTime;
     config.colorAtSwitchOnDayTime = config.colorAtSwitchOnDayTime.replace("geen", "green");
     config.colorAtSwitchOnNightTime = config.colorAtSwitchOnNightTime.replace("geen", "green");
-    config.dimSpeed = config.dimSpeed === undefined ? 5000 : Number(config.dimSpeed);
+    config.dimSpeed = (config.dimSpeed === undefined || config.dimSpeed === '') ? 5000 : Number(config.dimSpeed);
+    config.HSVDimSpeed = (config.HSVDimSpeed === undefined || config.HSVDimSpeed === '') ? 5000 : Number(config.HSVDimSpeed);
     config.invertDimTunableWhiteDirection = config.invertDimTunableWhiteDirection === undefined ? false : true;
+    node.HSVObject = null; //{ h, s, v };// Store the current light calculated HSV
 
     // Transform HEX in RGB and stringified json in json oblects.
     if (config.colorAtSwitchOnDayTime.indexOf("#") !== -1) {
@@ -245,6 +246,24 @@ module.exports = function (RED) {
               // { decr_incr: 0, data: 1 } : Start decreasing until { decr_incr: 0, data: 0 } is received.
               msg.payload = dptlib.fromBuffer(msg.knx.rawValue, dptlib.resolve(config.dptLightDIM));
               node.hueDimming(msg.payload.decr_incr, msg.payload.data, config.dimSpeed);
+              node.setNodeStatusHue({
+                fill: "green", shape: "dot", text: "KNX->HUE", payload: JSON.stringify(msg.payload),
+              });
+              break;
+            case config.GALightHSV_H_DIM:
+              // { decr_incr: 1, data: 1 } : Start increasing until { decr_incr: 0, data: 0 } is received.
+              // { decr_incr: 0, data: 1 } : Start decreasing until { decr_incr: 0, data: 0 } is received.
+              msg.payload = dptlib.fromBuffer(msg.knx.rawValue, dptlib.resolve(config.dptLightDIM));
+              node.hueDimmingHSV_H(msg.payload.decr_incr, msg.payload.data, config.HSVDimSpeed);
+              node.setNodeStatusHue({
+                fill: "green", shape: "dot", text: "KNX->HUE", payload: JSON.stringify(msg.payload),
+              });
+              break;
+            case config.GALightHSV_S_DIM:
+              // { decr_incr: 1, data: 1 } : Start increasing until { decr_incr: 0, data: 0 } is received.
+              // { decr_incr: 0, data: 1 } : Start decreasing until { decr_incr: 0, data: 0 } is received.
+              msg.payload = dptlib.fromBuffer(msg.knx.rawValue, dptlib.resolve(config.dptLightDIM));
+              node.hueDimmingHSV_S(msg.payload.decr_incr, msg.payload.data, config.HSVDimSpeed);
               node.setNodeStatusHue({
                 fill: "green", shape: "dot", text: "KNX->HUE", payload: JSON.stringify(msg.payload),
               });
@@ -513,6 +532,7 @@ module.exports = function (RED) {
           // STOP DIM
           if (node.timerStepDim !== undefined) clearInterval(node.timerStepDim);
           node.brightnessStep = null;
+          node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
           return;
         }
 
@@ -548,7 +568,7 @@ module.exports = function (RED) {
             node.updateKNXBrightnessState(node.brightnessStep); // Unnecessary, but necessary to set the KNX Status in real time.
             node.brightnessStep += numStep;
             if (node.brightnessStep > maxDimLevelLight) node.brightnessStep = maxDimLevelLight;
-            hueTelegram = { dimming: { brightness: node.brightnessStep }, dynamics: { duration: _dimSpeedInMillisecs + 300 } }; // + 100 is to avoid ladder effect
+            hueTelegram = { dimming: { brightness: node.brightnessStep }, dynamics: { duration: _dimSpeedInMillisecs + 500 } }; // + is to avoid ladder effect
             // Switch on the light if off
             if (node.currentHUEDevice.on !== undefined && node.currentHUEDevice.on.on === false) {
               hueTelegram.on = { on: true };
@@ -565,7 +585,7 @@ module.exports = function (RED) {
             node.updateKNXBrightnessState(node.brightnessStep); // Unnecessary, but necessary to set the KNX Status in real time.
             node.brightnessStep -= numStep;
             if (node.brightnessStep < minDimLevelLight) node.brightnessStep = minDimLevelLight;
-            hueTelegram = { dimming: { brightness: node.brightnessStep }, dynamics: { duration: _dimSpeedInMillisecs + 300 } };// + 100 is to avoid ladder effect
+            hueTelegram = { dimming: { brightness: node.brightnessStep }, dynamics: { duration: _dimSpeedInMillisecs + 500 } };// + 100 is to avoid ladder effect
             // Switch off the light if on
             if (node.currentHUEDevice.on !== undefined && node.currentHUEDevice.on.on === true && node.brightnessStep === 0) {
               hueTelegram.on = { on: false };
@@ -581,7 +601,7 @@ module.exports = function (RED) {
     // Start dimming tunable white
     // mirek: required(integer minimum: 153, maximum: 500)
     // ***********************************************************
-    node.hueDimmingTunableWhite = function hueDimming(_KNXaction, _KNXbrightness_DirectionTunableWhite, _dimSpeedInMillisecsTunableWhite = undefined) {
+    node.hueDimmingTunableWhite = function hueDimmingTunableWhite(_KNXaction, _KNXbrightness_DirectionTunableWhite, _dimSpeedInMillisecsTunableWhite = undefined) {
       // 23/23/2023 after many attempts to use dimming_delta function of the HueApeV2, loosing days of my life, without a decent success, will use the standard dimming calculations
       // i decide to go to the "step brightness" way.
       try {
@@ -590,6 +610,7 @@ module.exports = function (RED) {
           // STOP DIM
           if (node.timerStepDimTunableWhite !== undefined) clearInterval(node.timerStepDimTunableWhite);
           node.brightnessStepTunableWhite = null;
+          node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
           return;
         }
 
@@ -645,6 +666,183 @@ module.exports = function (RED) {
           }, _dimSpeedInMillisecsTunableWhite);
         }
       } catch (error) { }
+    };
+    // ***********************************************************
+
+    /**
+  * Starts dimming / stop dimming of the HUE
+  * @param {number} _KNXaction 1 for start, 0 for stop
+  * @param {number} _KNXbrightness_DirectionHSV_H 1 for up, 0 for down
+  * @param {number} _dimSpeedInMillisecsHSV Speed time in milliseconds
+  * @returns {}
+  */
+    node.hueDimmingHSV_H = function hueDimmingHSV_H(_KNXaction, _KNXbrightness_DirectionHSV_H, _dimSpeedInMillisecsHSV = undefined) {
+      // After many attempts to use dimming_delta function of the HueApiV2, loosing days of my life, without a decent success, will use the standard dimming calculations
+      // i decide to go to the "step brightness" way.
+      try {
+        if (_KNXbrightness_DirectionHSV_H === 0 && _KNXaction === 0) {
+          // STOP DIM
+          if (node.timerStepDimHSV_H !== undefined) clearInterval(node.timerStepDimHSV_H);
+          node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
+          return;
+        }
+
+        let xyBrightnessToHsv;
+        if (node.currentHUEDevice.color !== undefined && node.currentHUEDevice.color.xy !== undefined) {
+          // Get the XY + brightness from the lamp and transform it into HSV
+          // xyBrightnessToHsv = hueColorConverter.ColorConverter.xyBrightnessToHsv(node.currentHUEDevice.color.xy.x, node.currentHUEDevice.color.xy.y, node.currentHUEDevice.dimming.brightness);
+          xyBrightnessToHsv = hueColorConverter.ColorConverter.xyBrightnessToHsv(hueColorConverter.ColorConverter.scale(node.currentHUEDevice.color.xy.x, [0, 1], [0, 100]), hueColorConverter.ColorConverter.scale(node.currentHUEDevice.color.xy.y, [0, 1], [0, 100]), node.currentHUEDevice.dimming.brightness);
+        } else {
+          return; // Something wrong here.
+        }
+        let numStepHSV_H = 10; // Steps from 0 to 100
+
+        // Set the actual brightness to start with
+        if (node.brightnessStepHSV_H === null || node.brightnessStepHSV_H === undefined) node.brightnessStepHSV_H = node.currentHUEDevice.dimming.brightness;
+        node.brightnessStepHSV_H = Math.ceil(Number(node.brightnessStepHSV_H));
+
+        const minDimLevelLightHSV_H = 1;
+        const maxDimLevelLightHSV_H = 100;
+        numStepHSV_H = Math.round((maxDimLevelLightHSV_H - minDimLevelLightHSV_H) / numStepHSV_H, 0);
+        // Set the speed
+        _dimSpeedInMillisecsHSV = Math.ceil(_dimSpeedInMillisecsHSV / (maxDimLevelLightHSV_H / numStepHSV_H));
+
+        if (_KNXbrightness_DirectionHSV_H > 0 && _KNXaction === 1) {
+          // DIM UP
+          if (node.timerStepDimHSV_H !== undefined) clearInterval(node.timerStepDimHSV_H);
+          node.timerStepDimHSV_H = setInterval(() => {
+            node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
+            node.brightnessStepHSV_H += numStepHSV_H;
+            if (node.brightnessStepHSV_H > maxDimLevelLightHSV_H) node.brightnessStepHSV_H = maxDimLevelLightHSV_H;
+            //node.updateKNXLightHSV_H_State(node.brightnessStepHSV_H); // Unnecessary, but necessary to set the KNX Status in real time.
+            xyBrightnessToHsv.h = node.brightnessStepHSV_H;
+            const hsvToXYBri = hueColorConverter.ColorConverter.hsvToxyBrightness(xyBrightnessToHsv, node.currentHUEDevice.color.gamut);
+            const hueTelegram = { color: { xy: { x: hsvToXYBri.x, y: hsvToXYBri.y } }, dynamics: { duration: _dimSpeedInMillisecsHSV + 500 } }; // Add some ms, to prevent the "ladder" step effect.
+            // Switch on the light if off
+            if (node.currentHUEDevice.on !== undefined && node.currentHUEDevice.on.on === false) {
+              hueTelegram.on = { on: true };
+            }
+            node.currentHUEDevice.color.xy = hsvToXYBri;// Unnecessary, but necessary to set the KNX Status in real time.
+            node.updateKNXLightColorState(node.currentHUEDevice.color);
+            node.serverHue.hueManager.writeHueQueueAdd(node.hueDevice, hueTelegram, node.isGrouped_light === false ? "setLight" : "setGroupedLight");
+            if (node.brightnessStepHSV_H >= maxDimLevelLightHSV_H) {
+              if (node.timerStepDimHSV_H !== undefined) clearInterval(node.timerStepDimHSV_H);
+            }
+          }, _dimSpeedInMillisecsHSV);
+        }
+        if (_KNXbrightness_DirectionHSV_H > 0 && _KNXaction === 0) {
+          if (node.currentHUEDevice.on.on === false) return; // Don't dim down, if the light is already off.
+          // DIM DOWN
+          if (node.timerStepDimHSV_H !== undefined) clearInterval(node.timerStepDimHSV_H);
+          node.timerStepDimHSV_H = setInterval(() => {
+            node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
+            node.brightnessStepHSV_H -= numStepHSV_H; // *2 to speed up the things
+            if (node.brightnessStepHSV_H < minDimLevelLightHSV_H) node.brightnessStepHSV_H = minDimLevelLightHSV_H;
+            //node.updateKNXLightHSV_H_State(node.brightnessStepHSV_H); // Unnecessary, but necessary to set the KNX Status in real time.
+            // Produco l'RGB
+            xyBrightnessToHsv.h = node.brightnessStepHSV_H;
+            const hsvToXYBri = hueColorConverter.ColorConverter.hsvToxyBrightness(xyBrightnessToHsv, node.currentHUEDevice.color.gamut);
+            const hueTelegram = { color: { xy: { x: hsvToXYBri.x, y: hsvToXYBri.y } }, dynamics: { duration: _dimSpeedInMillisecsHSV + 500 } };
+            // Switch off the light if on
+            if (node.currentHUEDevice.on !== undefined && node.currentHUEDevice.on.on === true && node.brightnessStepHSV_H === 0) {
+              hueTelegram.on = { on: false };
+            }
+            node.currentHUEDevice.color.xy = hsvToXYBri;// Unnecessary, but necessary to set the KNX Status in real time.
+            node.updateKNXLightColorState(node.currentHUEDevice.color);
+            node.serverHue.hueManager.writeHueQueueAdd(node.hueDevice, hueTelegram, node.isGrouped_light === false ? "setLight" : "setGroupedLight");
+            if (node.brightnessStepHSV_H <= minDimLevelLightHSV_H) {
+              if (node.timerStepDimHSV_H !== undefined) clearInterval(node.timerStepDimHSV_H);
+            }
+          }, _dimSpeedInMillisecsHSV);
+        }
+      } catch (error) { /* empty */ }
+    };
+    /**
+  * Starts dimming / stop dimming of the Saturation
+  * @param {number} _KNXaction 1 for start, 0 for stop
+  * @param {number} _KNXbrightness_DirectionHSV_S 1 for up, 0 for down
+  * @param {number} _dimSpeedInMillisecsHSV Speed time in milliseconds
+  * @returns {}
+  */
+    node.hueDimmingHSV_S = function hueDimmingHSV_S(_KNXaction, _KNXbrightness_DirectionHSV_S, _dimSpeedInMillisecsHSV = undefined) {
+      // After many attempts to use dimming_delta function of the HueApiV2, loosing days of my life, without a decent success, will use the standard dimming calculations
+      // i decide to go to the "step brightness" way.
+      try {
+        if (_KNXbrightness_DirectionHSV_S === 0 && _KNXaction === 0) {
+          // STOP DIM
+          if (node.timerStepDimHSV_S !== undefined) clearInterval(node.timerStepDimHSV_S);
+          node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
+          return;
+        }
+
+        let xyBrightnessToHsv;
+        if (node.currentHUEDevice.color !== undefined && node.currentHUEDevice.color.xy !== undefined) {
+          // Get the XY + brightness from the lamp and transform it into HSV
+          xyBrightnessToHsv = hueColorConverter.ColorConverter.xyBrightnessToHsv(hueColorConverter.ColorConverter.scale(node.currentHUEDevice.color.xy.x, [0, 1], [0, 100]), hueColorConverter.ColorConverter.scale(node.currentHUEDevice.color.xy.y, [0, 1], [0, 100]), node.currentHUEDevice.dimming.brightness);
+        } else {
+          return; // Something wrong here.
+        }
+        let numStepHSV_S = 10; // Steps from 0 to 100
+
+        // Set the actual brightness to start with
+        if (node.brightnessStepHSV_S === null || node.brightnessStepHSV_S === undefined) node.brightnessStepHSV_S = node.currentHUEDevice.dimming.brightness;
+        node.brightnessStepHSV_S = Math.ceil(Number(node.brightnessStepHSV_S));
+
+        const minDimLevelLightHSV_S = 1;
+        const maxDimLevelLightHSV_S = 100;
+        numStepHSV_S = Math.round((maxDimLevelLightHSV_S - minDimLevelLightHSV_S) / numStepHSV_S, 0);
+        // Set the speed
+        _dimSpeedInMillisecsHSV = Math.ceil(_dimSpeedInMillisecsHSV / (maxDimLevelLightHSV_S / numStepHSV_S));
+
+        if (_KNXbrightness_DirectionHSV_S > 0 && _KNXaction === 1) {
+          // DIM UP
+          if (node.timerStepDimHSV_S !== undefined) clearInterval(node.timerStepDimHSV_S);
+          node.timerStepDimHSV_S = setInterval(() => {
+            node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
+            node.brightnessStepHSV_S += numStepHSV_S;
+            if (node.brightnessStepHSV_S > maxDimLevelLightHSV_S) node.brightnessStepHSV_S = maxDimLevelLightHSV_S;
+            //node.updateKNXLightHSV_S_State(node.brightnessStepHSV_S); // Unnecessary, but necessary to set the KNX Status in real time.
+            xyBrightnessToHsv.s = node.brightnessStepHSV_S;
+            const hsvToXYBri = hueColorConverter.ColorConverter.hsvToxyBrightness(xyBrightnessToHsv, node.currentHUEDevice.color.gamut);
+            const hueTelegram = { color: { xy: { x: hsvToXYBri.x, y: hsvToXYBri.y } }, dynamics: { duration: _dimSpeedInMillisecsHSV + 500 } }; // Add some ms, to prevent the "ladder" step effect.
+            // Switch on the light if off
+            if (node.currentHUEDevice.on !== undefined && node.currentHUEDevice.on.on === false) {
+              hueTelegram.on = { on: true };
+            }
+            node.currentHUEDevice.color.xy = cloneDeep(hsvToXYBri);// Unnecessary, but necessary to set the KNX Status in real time.
+            node.updateKNXLightColorState(node.currentHUEDevice.color);
+            node.serverHue.hueManager.writeHueQueueAdd(node.hueDevice, hueTelegram, node.isGrouped_light === false ? "setLight" : "setGroupedLight");
+            if (node.brightnessStepHSV_S >= maxDimLevelLightHSV_S) {
+              if (node.timerStepDimHSV_S !== undefined) clearInterval(node.timerStepDimHSV_S);
+            }
+          }, _dimSpeedInMillisecsHSV);
+        }
+        if (_KNXbrightness_DirectionHSV_S > 0 && _KNXaction === 0) {
+          if (node.currentHUEDevice.on.on === false) return; // Don't dim down, if the light is already off.
+          // DIM DOWN
+          if (node.timerStepDimHSV_S !== undefined) clearInterval(node.timerStepDimHSV_S);
+          node.timerStepDimHSV_S = setInterval(() => {
+            node.serverHue.hueManager.deleteHueQueue(node.hueDevice); // Clear dimming queue.
+            node.brightnessStepHSV_S -= numStepHSV_S; // *2 to speed up the things
+            if (node.brightnessStepHSV_S < minDimLevelLightHSV_S) node.brightnessStepHSV_S = minDimLevelLightHSV_S;
+            //node.updateKNXLightHSV_S_State(node.brightnessStepHSV_S); // Unnecessary, but necessary to set the KNX Status in real time.
+            // Produco l'RGB
+            xyBrightnessToHsv.s = node.brightnessStepHSV_S;
+            const hsvToXYBri = hueColorConverter.ColorConverter.hsvToxyBrightness(xyBrightnessToHsv, node.currentHUEDevice.color.gamut);
+            const hueTelegram = { color: { xy: { x: hsvToXYBri.x, y: hsvToXYBri.y } }, dynamics: { duration: _dimSpeedInMillisecsHSV + 500 } };
+            // Switch off the light if on
+            if (node.currentHUEDevice.on !== undefined && node.currentHUEDevice.on.on === true && node.brightnessStepHSV_S === 0) {
+              hueTelegram.on = { on: false };
+            }
+            node.currentHUEDevice.color.xy = cloneDeep(hsvToXYBri);// Unnecessary, but necessary to set the KNX Status in real time.
+            node.updateKNXLightColorState(node.currentHUEDevice.color);
+            node.serverHue.hueManager.writeHueQueueAdd(node.hueDevice, hueTelegram, node.isGrouped_light === false ? "setLight" : "setGroupedLight");
+            if (node.brightnessStepHSV_S <= minDimLevelLightHSV_S) {
+              if (node.timerStepDimHSV_S !== undefined) clearInterval(node.timerStepDimHSV_S);
+            }
+          }, _dimSpeedInMillisecsHSV);
+        }
+      } catch (error) { /* empty */ }
     };
     // ***********************************************************
 
@@ -835,6 +1033,82 @@ module.exports = function (RED) {
       }
     };
 
+    //     /**
+    //  * Update the internal HSV variable status and send the valuge in % to the KNX BUS
+    //  * @param {number} _value 0-100%
+    //  * @param {string} _outputtype "write" is the default KNX command
+    //  * * @returns {}
+    //  */
+    //     node.updateKNXLightHSV_H_State = function updateKNXLightHSV_H_State(_value, _outputtype = "write") {
+    //       if (config.GALightHSV_H_State !== undefined && config.GALightHSV_H_State !== "") {
+    //         const knxMsgPayload = {};
+    //         knxMsgPayload.topic = config.GALightHSV_H_State;
+    //         knxMsgPayload.dpt = config.dptLightHSV_H_State;
+    //         if (config.dptLightHSV_H_State === "5.001") {
+    //           knxMsgPayload.payload = 100 - _value;
+    //         }
+    //         // Send to KNX bus
+    //         if (knxMsgPayload.topic !== "" && knxMsgPayload.topic !== undefined) {
+    //           if (node.server !== null && node.server !== undefined) {
+    //             node.server.writeQueueAdd({
+    //               grpaddr: knxMsgPayload.topic,
+    //               payload: knxMsgPayload.payload,
+    //               dpt: knxMsgPayload.dpt,
+    //               outputtype: _outputtype || "write",
+    //               nodecallerid: node.id,
+    //             });
+    //           }
+    //         }
+    //         node.setNodeStatusHue({
+    //           fill: "blue",
+    //           shape: "ring",
+    //           text: "HUE->KNX HSV (H)",
+    //           payload: knxMsgPayload.payload,
+    //         });
+    //       }
+    //     };
+
+    //     /**
+    //  * Update the internal HSV variable status and send the valuge in % to the KNX BUS
+    //  * @param {number} _value 0-100%
+    //  * @param {string} _outputtype "write" is the default KNX command
+    //  * @returns {}
+    //  */
+    //     node.updateKNXLightHSV_S_State = function updateKNXLightHSV_S_State(_value, _outputtype = "write") {
+    //       if (config.GALightHSV_S_State !== undefined && config.GALightHSV_S_State !== "") {
+    //         const knxMsgPayload = {};
+    //         knxMsgPayload.topic = config.GALightHSV_S_State;
+    //         knxMsgPayload.dpt = config.dptLightHSV_S_State;
+    //         if (config.dptLightHSV_S_State === "5.001") {
+    //           knxMsgPayload.payload = _value;
+    //         }
+    //         // Send to KNX bus
+    //         if (knxMsgPayload.topic !== "" && knxMsgPayload.topic !== undefined) {
+    //           if (node.server !== null && node.server !== undefined) {
+    //             node.server.writeQueueAdd({
+    //               grpaddr: knxMsgPayload.topic,
+    //               payload: knxMsgPayload.payload,
+    //               dpt: knxMsgPayload.dpt,
+    //               outputtype: _outputtype || "write",
+    //               nodecallerid: node.id,
+    //             });
+    //           }
+    //         }
+    //         node.setNodeStatusHue({
+    //           fill: "blue",
+    //           shape: "ring",
+    //           text: "HUE->KNX HSV (H)",
+    //           payload: knxMsgPayload.payload,
+    //         });
+    //       }
+    //     };
+
+    /**
+* Update the KNC colors and HSV states group addresses
+* @param {object} _value {xy:{x,y}} in 0-1 scale
+* @param {string} _outputtype "write" is the default KNX command
+* @returns {}
+*/
     node.updateKNXLightColorState = function updateKNXLightColorState(_value, _outputtype = "write") {
       if (config.GALightColorState !== undefined && config.GALightColorState !== "") {
         if (_value.xy === undefined || _value.xy.x === undefined) return;
@@ -867,6 +1141,60 @@ module.exports = function (RED) {
             payload: knxMsgPayload.payload,
           });
         } catch (error) { }
+      }
+      if (config.GALightHSV_S_State !== undefined && config.GALightHSV_S_State !== "") {
+        const knxMsgPayload = {};
+        knxMsgPayload.topic = config.GALightHSV_S_State;
+        knxMsgPayload.dpt = config.dptLightHSV_S_State;
+        if (config.dptLightHSV_S_State === "5.001") {
+          knxMsgPayload.payload = hueColorConverter.ColorConverter.xyBrightnessToHsv(_value.xy.x, _value.xy.y, node.currentHUEDevice !== undefined && node.currentHUEDevice.dimming !== undefined && node.currentHUEDevice.dimming.brightness === undefined ? node.currentHUEDevice.dimming.brightness : 100, false);
+          knxMsgPayload.payload = knxMsgPayload.payload.s;
+        }
+        // Send to KNX bus
+        if (knxMsgPayload.topic !== "" && knxMsgPayload.topic !== undefined) {
+          if (node.server !== null && node.server !== undefined) {
+            node.server.writeQueueAdd({
+              grpaddr: knxMsgPayload.topic,
+              payload: knxMsgPayload.payload,
+              dpt: knxMsgPayload.dpt,
+              outputtype: _outputtype || "write",
+              nodecallerid: node.id,
+            });
+          }
+        }
+        node.setNodeStatusHue({
+          fill: "blue",
+          shape: "ring",
+          text: "HUE->KNX HSV (S)",
+          payload: knxMsgPayload.payload,
+        });
+      }
+      if (config.GALightHSV_H_State !== undefined && config.GALightHSV_H_State !== "") {
+        const knxMsgPayload = {};
+        knxMsgPayload.topic = config.GALightHSV_H_State;
+        knxMsgPayload.dpt = config.dptLightHSV_H_State;
+        if (config.dptLightHSV_H_State === "5.001") {
+          knxMsgPayload.payload = hueColorConverter.ColorConverter.xyBrightnessToHsv(_value.xy.x, _value.xy.y, node.currentHUEDevice !== undefined && node.currentHUEDevice.dimming !== undefined && node.currentHUEDevice.dimming.brightness === undefined ? node.currentHUEDevice.dimming.brightness : 100, false);
+          knxMsgPayload.payload = 100 - knxMsgPayload.payload.h;
+        }
+        // Send to KNX bus
+        if (knxMsgPayload.topic !== "" && knxMsgPayload.topic !== undefined) {
+          if (node.server !== null && node.server !== undefined) {
+            node.server.writeQueueAdd({
+              grpaddr: knxMsgPayload.topic,
+              payload: knxMsgPayload.payload,
+              dpt: knxMsgPayload.dpt,
+              outputtype: _outputtype || "write",
+              nodecallerid: node.id,
+            });
+          }
+        }
+        node.setNodeStatusHue({
+          fill: "blue",
+          shape: "ring",
+          text: "HUE->KNX HSV (H)",
+          payload: knxMsgPayload.payload,
+        });
       }
     };